\name{FLOWMR-package}
\alias{FLOWMR-package}
\alias{FLOWMR}
\docType{package}
\title{
\packageTitle{FLOWMR}
}
\description{
\packageDescription{FLOWMR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FLOWMR}
\packageIndices{FLOWMR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FLOWMR}

Maintainer: \packageMaintainer{FLOWMR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\examples{
# simple examples of the most important functions ~~
}
