% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{get_indirect_samples}
\alias{get_indirect_samples}
\title{Posterior samples of indirect effects}
\usage{
get_indirect_samples(results, K, warmup = 3000, path = NULL, ind = NULL)
}
\arguments{
\item{results}{results from gibbs_wrapper}

\item{K}{number of phenotypes}

\item{warmup}{The length of warmup period. Default is 3000}

\item{path}{A vector representing the pathway of interests}

\item{ind}{index of indirect effect of interest if path is NULL}
}
\value{
A vector of posterior samples
}
\description{
Estimate the indirrect effect of an exposure on the outcome
}
