% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{get_posterior_samples}
\alias{get_posterior_samples}
\title{Posterior samples}
\usage{
get_posterior_samples(raw, par, K, ind = NULL, warmup = 3000)
}
\arguments{
\item{raw}{raw results}

\item{par}{parameter of interets}

\item{K}{number of traits}

\item{ind}{index of the parameter}

\item{warump}{The length of warmup period. Default is 3000}
}
\value{
A vector of posterior samples from all chains
}
\description{
Estimate the total effect of an exposure on the outcome
}
