% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{get_total_samples}
\alias{get_total_samples}
\title{Posterior samples of total effects}
\usage{
get_total_samples(results, K, ind, warmup = 3000)
}
\arguments{
\item{results}{results from gibbs_wrapper}

\item{K}{number of phenotypes}

\item{ind}{index of interet}

\item{warump}{The length of warmup period. Default is 3000}
}
\value{
A K by K by 4 array of summary of total effect.
}
\description{
Estimate the total effect of an exposure on the outcome
}
